13.3

+++ dtrace/lib/libctf/common/ctf_hash.c
@@ -27,12 +27,12 @@
 
 #include <ctf_impl.h>
 
-static const uint16_t _CTF_EMPTY[1] = { 0 };
+static const uint32_t _CTF_EMPTY[1] = { 0 };
 
 int
 ctf_hash_create(ctf_hash_t *hp, unsigned long nelems)
 {
-	if (nelems > USHRT_MAX)
+	if (nelems > UINT32_MAX)
 		return (EOVERFLOW);
 
 	/*
@@ -41,7 +41,7 @@ ctf_hash_create(ctf_hash_t *hp, unsigned long nelems)
 	 */
 	if (nelems == 0) {
 		bzero(hp, sizeof (ctf_hash_t));
-		hp->h_buckets = (uint16_t *)_CTF_EMPTY;
+		hp->h_buckets = (uint32_t *)_CTF_EMPTY;
 		hp->h_nbuckets = 1;
 		return (0);
 	}
@@ -50,7 +50,7 @@ ctf_hash_create(ctf_hash_t *hp, unsigned long nelems)
 	hp->h_nelems = nelems + 1;	/* we use index zero as a sentinel */
 	hp->h_free = 1;			/* first free element is index 1 */
 
-	hp->h_buckets = ctf_alloc(sizeof (uint16_t) * hp->h_nbuckets);
+	hp->h_buckets = ctf_alloc(sizeof (uint32_t) * hp->h_nbuckets);
 	hp->h_chains = ctf_alloc(sizeof (ctf_helem_t) * hp->h_nelems);
 
 	if (hp->h_buckets == NULL || hp->h_chains == NULL) {
@@ -58,7 +58,7 @@ ctf_hash_create(ctf_hash_t *hp, unsigned long nelems)
 		return (EAGAIN);
 	}
 
-	bzero(hp->h_buckets, sizeof (uint16_t) * hp->h_nbuckets);
+	bzero(hp->h_buckets, sizeof (uint32_t) * hp->h_nbuckets);
 	bzero(hp->h_chains, sizeof (ctf_helem_t) * hp->h_nelems);
 
 	return (0);
@@ -90,7 +90,7 @@ ctf_hash_compute(const char *key, size_t len)
 }
 
 int
-ctf_hash_insert(ctf_hash_t *hp, ctf_file_t *fp, uint16_t type, uint32_t name)
+ctf_hash_insert(ctf_hash_t *hp, ctf_file_t *fp, uint32_t type, uint32_t name)
 {
 	ctf_strs_t *ctsp = &fp->ctf_str[CTF_NAME_STID(name)];
 	const char *str = ctsp->cts_strs + CTF_NAME_OFFSET(name);
@@ -127,7 +127,7 @@ ctf_hash_insert(ctf_hash_t *hp, ctf_file_t *fp, uint16_t type, uint32_t name)
  * If the key is not present, then call ctf_hash_insert() and hash it in.
  */
 int
-ctf_hash_define(ctf_hash_t *hp, ctf_file_t *fp, uint16_t type, uint32_t name)
+ctf_hash_define(ctf_hash_t *hp, ctf_file_t *fp, uint32_t type, uint32_t name)
 {
 	const char *str = ctf_strptr(fp, name);
 	ctf_helem_t *hep = ctf_hash_lookup(hp, fp, str, strlen(str));
@@ -145,7 +145,7 @@ ctf_hash_lookup(ctf_hash_t *hp, ctf_file_t *fp, const char *key, size_t len)
 	ctf_helem_t *hep;
 	ctf_strs_t *ctsp;
 	const char *str;
-	uint16_t i;
+	uint32_t i;
 
 	unsigned long h = ctf_hash_compute(key, len) % hp->h_nbuckets;
 
@@ -165,7 +165,7 @@ void
 ctf_hash_destroy(ctf_hash_t *hp)
 {
 	if (hp->h_buckets != NULL && hp->h_nbuckets != 1) {
-		ctf_free(hp->h_buckets, sizeof (uint16_t) * hp->h_nbuckets);
+		ctf_free(hp->h_buckets, sizeof (uint32_t) * hp->h_nbuckets);
 		hp->h_buckets = NULL;
 	}
